Simplify navigation blocker implementation
